* Moved the CalibrationError to the package exceptions.py
