+++ qiskit_experiments/calibration/drag.py
@@ -23,7 +23,7 @@ import qiskit.pulse as pulse
 from qiskit.providers.options import Options
 
 from qiskit_experiments.base_experiment import BaseExperiment
-from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.exceptions import CalibrationError
 from qiskit_experiments.calibration.analysis.drag_analysis import DragCalAnalysis
