+++ qiskit_experiments/calibration/management/backend_calibrations.py
@@ -20,7 +20,7 @@ import copy
 from qiskit.providers.backend import BackendV1 as Backend
 from qiskit.circuit import Parameter
 from qiskit_experiments.calibration.management.calibrations import Calibrations, ParameterKey
-from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.exceptions import CalibrationError
 
 
 class FrequencyElement(Enum):
