+++ qiskit_experiments/calibration/management/calibrations.py
@@ -35,7 +35,7 @@ from qiskit.pulse import (
 )
 from qiskit.pulse.channels import PulseChannel
 from qiskit.circuit import Parameter, ParameterExpression
-from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.exceptions import CalibrationError
 from qiskit_experiments.calibration.management.parameter_value import ParameterValue
 from qiskit_experiments.calibration.management.calibration_key_types import (
     ParameterKey,
