+++ test/calibration/test_calibrations.py
@@ -33,7 +33,7 @@ import qiskit.pulse as pulse
 from qiskit.test import QiskitTestCase
 from qiskit_experiments.calibration.management.calibrations import Calibrations, ParameterKey
 from qiskit_experiments.calibration.management.parameter_value import ParameterValue
-from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.exceptions import CalibrationError
 
 
 class TestCalibrationsBasic(QiskitTestCase):
