+++ test/calibration/test_update_library.py
@@ -25,7 +25,7 @@ from qiskit.test.mock import FakeAthens
 from qiskit_experiments.calibration.rabi import Rabi
 from qiskit_experiments.calibration.drag import DragCal
 from qiskit_experiments.calibration.management.calibrations import Calibrations
-from qiskit_experiments.calibration.management.exceptions import CalibrationError
+from qiskit_experiments.exceptions import CalibrationError
 from qiskit_experiments.calibration.management.update_library import Frequency, Amplitude, Drag
 from qiskit_experiments.calibration.management.backend_calibrations import BackendCalibrations
 from qiskit_experiments.characterization.qubit_spectroscopy import QubitSpectroscopy
